pyproject.toml
[tool.poetry]
name = "pyjiit"
version = "0.1.0"
description = "Python wrapper for JIIT webportal"
authors = ["Harsh Sharma <harsh@codelif.in>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
